--- src/components/dashboard/CategoriesCard/CategoriesCard.module.css
@@ -1,3 +1,0 @@
-.Wrapper {
-  background: radial-gradient(50% 50% at 50% 50%, #E8E0C6 0%, #C9D9DE 100%);
-}
