--- src/components/dashboard/DataCard/DataCard.module.css
@@ -1,3 +1,0 @@
-.Wrapper {
-  background: radial-gradient(50% 50% at 50% 50%, #B2D8E4 0%, #E1EDF2 100%);
-}
