+++ src/styles/globals.css
@@ -201,7 +201,7 @@ label {
 	background: radial-gradient(50% 50% at 50% 50%, rgba(255, 97, 220, 0.8) 0%, rgba(194, 200, 255, 0.8) 64.06%, rgba(225, 237, 242, 0.8) 100%);
 }
 
-.void-green-gradient {
+.dashboard-card-bg {
 	background: linear-gradient(221.46deg, rgba(2, 7, 10, 0) 0%, #050D11 100%);
 }
